Update samples with new reviews method usage
