+++ appvox-examples/src/main/kotlin/com/company/project/itunesrss/ExportItunesRssReviewsToCsv.kt
@@ -1,30 +1,23 @@
-package com.examples.review.googleplay
+package com.company.project.itunesrss
 
 import com.opencsv.CSVWriter
-import io.appvox.GooglePlay
+import io.appvox.ItunesRss
 import io.appvox.configuration.RequestConfiguration
 import io.appvox.exception.AppVoxException
-import io.appvox.review.googleplay.constant.GooglePlayLanguage
-import io.appvox.review.googleplay.constant.GooglePlaySortType
+import io.appvox.review.itunesrss.constant.AppStoreRegion
+import io.appvox.review.itunesrss.constant.ItunesRssSortType
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.take
 import kotlinx.coroutines.runBlocking
 import java.io.FileWriter
 import java.io.IOException
-import kotlin.contracts.ExperimentalContracts
 
-@ExperimentalContracts
 fun main() = runBlocking {
-    val appId = "com.twitter.android"
-    val sortType = GooglePlaySortType.RELEVANT
-    val reviewLanguage = GooglePlayLanguage.ENGLISH_US
+
+    val appId = "333903271"
     val maxReviewCount = 100
 
-    val fileName =
-        appId +
-            "_${sortType.name.toLowerCase()}" +
-            "_${reviewLanguage.name.toLowerCase()}" +
-            "_$maxReviewCount.csv"
+    val fileName = "${appId}_us_$maxReviewCount.csv"
     val fileWriter = FileWriter(fileName)
 
     try {
@@ -37,37 +30,35 @@ fun main() = runBlocking {
         )
         val columns: Array<String> =
             arrayOf(
-                "id", "rating", "userName", "userProfile",
-                "title", "comment", "commentTime", "appVersion",
-                "likeCount", "replyComment", "replyTime", "url"
+                "id", "rating", "userName",
+                "title", "appVersion", "comment", "commentTime",
+                "replyComment", "replyTime", "url"
             )
         csvWriter.writeNext(columns)
 
         val config = RequestConfiguration(
-//        proxy = Proxy(Proxy.Type.HTTP, InetSocketAddress("localhost", 8080)),
+//        proxy = Proxy(HTTP, InetSocketAddress("localhost", 8080)),
 //        proxyAuthentication = PasswordAuthentication("my-proxy-username", "my-proxy-password".toCharArray()),
-            delay = 5000
+            delay = 3000
         )
-        GooglePlay(config).reviews(
-            appId = appId,
-            sortType = sortType,
-            language = reviewLanguage,
-            batchSize = 40
-        )
+
+        val appStore = ItunesRss(config)
+        appStore.reviews {
+                this.appId = appId
+                region = AppStoreRegion.UNITED_STATES
+                sortType = ItunesRssSortType.RECENT
+        }
             .take(maxReviewCount)
             .collect { review ->
                 val csvReview: Array<String?> = arrayOf(
                     review.id,
                     review.latestUserComment.rating.toString(),
-                    review.latestUserComment.name,
-                    review.latestUserComment.avatar,
+                    review.latestUserComment.userName,
                     review.latestUserComment.title,
-                    review.latestUserComment.text,
-                    review.latestUserComment.lastUpdateTime.toString(),
                     review.latestUserComment.appVersion,
-                    review.latestUserComment.likeCount.toString(),
                     review.latestUserComment.text,
-                    review.latestUserComment.lastUpdateTime.toString(),
+                    review.latestUserComment.time.toString(),
+                    review.latestUserComment.likeCount.toString(),
                     review.url
                 )
                 csvWriter.writeNext(csvReview)
